+++ src/dpr/routes/asyncReports.ts
@@ -6,6 +6,7 @@ import * as AsyncReportUtils from '../utils/renderAsyncReport'
 
 import AsyncReportslistUtils from '../utils/asyncReportsUtils'
 import { Services } from '../types/Services'
+import logger from '../utils/logger'
 
 export default function routes({
   router,
@@ -19,11 +20,12 @@ export default function routes({
   templatePath?: string
 }) {
   const asyncErrorHandler: RequestHandler = async (req, res) => {
+    logger.error(`Error: ${JSON.stringify(req.body)}`)
     let { error } = req.body
 
-    if (error.message) {
+    if (error && error.message) {
       error = { userMessage: `${error.name}: ${error.message}`, status: 'FAILED', stack: error.stack }
-    } else if (error.data) {
+    } else if (error && error.data) {
       error = error.data
     }
 
@@ -131,6 +133,7 @@ export default function routes({
     } catch (error) {
       req.body.title = 'Failed to retrieve Report'
       req.body.errorDescription = 'We were unable to retrieve this report for the following reason:'
+      req.body.error = error
       next()
     }
   }