+++ src/dpr/types/Template.d.ts
@@ -1,1 +1,0 @@
-export type Template = 'list' | 'list-section' | 'list-aggregate' | 'list-tab' | 'crosstab' | 'summary'