+++ src/dpr/types/Templates.d.ts
@@ -1,0 +1,3 @@
+export type Template = 'list' | 'list-section' | 'list-aggregate' | 'list-tab' | 'crosstab' | 'summary'
+
+export type SummaryTemplate = 'table-header' | 'table-footer' | 'section-footer' | 'page-header' | 'page-footer'